fix: removed parameters from config, consolidated _helper function in to its own class called LLaMaStack, hidden_grow_factor switched from int to float in docstrings

+++ fms/modules/feedforward.py
@@ -75,7 +75,7 @@ class GatedLinearUnit(nn.Module):
     ----
     emb_dim : int
         Dimensionality of input and output vectors.
-    hidden_grow_factor : int
+    hidden_grow_factor : float
         Sets dimensionality of inner latent space (emb_dim * hidden_grow_factor)
     multiple_of : Optional[int]
         Ensure inner latent space is a multiple of this parameter if defined (useful for
